+++ libsolidity/interface/GasEstimator.cpp
@@ -45,8 +45,10 @@ GasEstimator::ASTGasConsumptionSelfAccumulated GasEstimator::structuralEstimatio
 	solAssert(std::count(_ast.begin(), _ast.end(), nullptr) == 0, "");
 	map<SourceLocation, GasConsumption> particularCosts;
 
-	ControlFlowGraph cfg(_items);
-	for (BasicBlock const& block: cfg.optimisedBlocks())
+	set<size_t> tagsPushedByParent;
+	ControlFlowGraph cfg(_items, tagsPushedByParent);
+	// Enable gathering of knowledge
+	for (BasicBlock const& block: cfg.optimisedBlocks(true))
 	{
 		assertThrow(!!block.startState, OptimizerException, "");
 		GasMeter meter(block.startState->copy());